requirements.txt
python-telegram-bot >= 20.2
cryptography >= 40.0.2
